changed all to uint

+++ exercises/gas/contracts/Gas.sol
@@ -2,27 +2,27 @@
 pragma solidity 0.8.17;
 
 contract GasContract {
-    uint16 public immutable totalSupply; // cannot be updated
+    uint public immutable totalSupply; // cannot be updated
     address[5] public administrators;
 
     struct Payment {
-        uint8 paymentType;
-        uint8 paymentID;
-        uint16 amount;
+        uint paymentType;
+        uint paymentID;
+        uint amount;
     }
     struct ImportantStruct {
-        uint8 valueA; // max 3 digits
-        uint48 bigValue;
-        uint8 valueB; // max 3 digits
+        uint valueA; // max 3 digits
+        uint bigValue;
+        uint valueB; // max 3 digits
     }
 
-    event Transfer(address recipient, uint16 amount);
+    event Transfer(address recipient, uint amount);
 
-    mapping(address => uint16) balances;
+    mapping(address => uint) balances;
     mapping(address => Payment[]) payments;
-    mapping(address => uint8) public whitelist;
+    mapping(address => uint) public whitelist;
 
-    constructor(address[5] memory _admins, uint16 _totalSupply) {
+    constructor(address[5] memory _admins, uint _totalSupply) {
         totalSupply = _totalSupply;
         administrators = _admins;
         balances[msg.sender] = totalSupply;
@@ -30,7 +30,7 @@ contract GasContract {
 
     function transfer(
         address _recipient,
-        uint16 _amount,
+        uint _amount,
         string calldata _name
     ) external { unchecked{
         balances[_recipient] += _amount;}
@@ -38,18 +38,18 @@ contract GasContract {
         payments[msg.sender].push(Payment(1, 0, _amount));
     }
 
-    function balanceOf(address _user) external view returns (uint16) {
+    function balanceOf(address _user) external view returns (uint) {
         return balances[_user];
     }
 
     function updatePayment(
         address _user,
-        uint8 _ID,
-        uint16 _amount,
-        uint8 _type
+        uint _ID,
+        uint _amount,
+        uint _type
     ) external {
         unchecked{
-        for (uint8 i = 0; i < 6; ++i) {
+        for (uint i = 0; i < 6; ++i) {
             if (administrators[i] == msg.sender) {
                 payments[_user][0].paymentType = _type;
                 payments[_user][0].amount = _amount;
@@ -71,13 +71,13 @@ contract GasContract {
         return true;
     }
 
-    function addToWhitelist(address _userAddrs, uint8 _tier) external {
+    function addToWhitelist(address _userAddrs, uint _tier) external {
         whitelist[_userAddrs] = _tier;
     }
 
     function whiteTransfer(
         address _recipient,
-        uint16 _amount,
+        uint _amount,
         ImportantStruct calldata _struct
     ) external { 
         unchecked {
